Ocr branch

+++ anuvaad-api/anuvaad-metrics/anuvaad-org-judgement-count/src/resources/scheduler_jobs.py
@@ -44,13 +44,14 @@ log_info("Mongo connected", MODULE_CONTEXT)
 # tz = IST()
 
 # @.scheduled_job("interval", id="get_data_from_db", hours=6)
-# @schedule_job.scheduled_job(
-#     "cron", id="my_job_id1", day_of_week="sun", hour="02", minute="00"
-# )
-def get_trans_user_data_from_db_weekly_crn():
+@schedule_job.scheduled_job(
+    "cron", id="my_job_id1", day_of_week="mon-fri", hour="00,06,12,18", minute="00"
+)
+def get_trans_user_data_from_db_cron():
     users = config.EMAIL_NOTIFIER
     log_info("fetch data started", MODULE_CONTEXT)
     # filename = uuid.uuid4().hex
+    stats_file = config.STATS_FILE
     weekly_cron_file_name1 = config.WEEKLY_CRON_FILE_NAME1
     weekly_cron_file_name2 = config.WEEKLY_CRON_FILE_NAME2
     daily_cron_file_name1 = config.DAILY_CRON_FILE_NAME1
@@ -58,11 +59,13 @@ def get_trans_user_data_from_db_weekly_crn():
     # file_save = str(filename)[:-10]+'_USER_WISE_JUD_Org_level_Statistics.csv'
     if os.path.exists(
         config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name1
-    ) and os.path.exists(config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name2):
+    ) and os.path.exists(config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name2 
+    ) and os.path.exists(config.DOWNLOAD_FOLDER + "/" + stats_file):
         os.remove(config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name1)
         os.remove(config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name2)
-        os.remove(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name1)
-        os.remove(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name2)
+        os.remove(config.DOWNLOAD_FOLDER + "/" + stats_file)
+        # os.remove(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name1)
+        # os.remove(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name2)
     else:
         msg = generate_email_notification(
             users, "could not get the data files not found"
@@ -70,75 +73,44 @@ def get_trans_user_data_from_db_weekly_crn():
         send_email(msg)
         log_info(f"Generated alert email scheduler files not found ", MODULE_CONTEXT)
     user_docs = stats.get_all_users_active_inactive(usr_collection)
+    user_df =pd.json_normalize(user_docs)
     log_info(
         f"Data returned from user {config.USER_COLLECTION} collection", MODULE_CONTEXT
     )
     try:
         from_date, end_date = stats.get_time_frame_for_analytics()
-        done = 0
-        while True:
-            for doc in user_docs:
-                # log_info(f'fetching details for {doc} userID',MODULE_CONTEXT)
-                
-                try:
-                    
 
-                        ch_docs = (
-                            stats.fetch_data_for_language_trans_tokenized_for_scheduer_only(
-                                ch_collection, doc, from_date, end_date
-                            )
-                        )
-                        saved_docs = stats.fetch_data_for_userwise_trans_user_tokenized(
-                            ch_collection, doc, from_date, end_date
-                        )
-                        # log_info(f'Details collected for for userID : {doc} ',MODULE_CONTEXT)
-                        write_to_csv_user(
-                            [x for x in ch_docs],
-                            (config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name1),
-                        )
-                        write_to_csv_user(
-                            [x for x in saved_docs],
-                            (config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name2),
-                        )
-                except Exception as e:
-                    log_exception(
-                        "error in fetching the data : {}".format(str(e)),
-                        MODULE_CONTEXT,
-                        e,
-                    )
-            done = 1
-            if done == 1:
-                break
-            log_info(
-                f"Data written into files {weekly_cron_file_name1,weekly_cron_file_name2}",
-                MODULE_CONTEXT,
+        ch_docs = (
+            stats.fetch_data_for_language_trans_tokenized_for_scheduer_only(
+                ch_collection, from_date, end_date
             )
-        if not os.path.exists(
-            config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name1
-        ) and not os.path.exists(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name2):
-            shutil.copyfile(
-                config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name1,
-                config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name1,
-            )
-            shutil.copyfile(
-                config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name2,
-                config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name2,
-            )
-            log_info(
-                f"files copied from weekly cron to daily cron {daily_cron_file_name1,daily_cron_file_name2}",
-                MODULE_CONTEXT,
-            )
-            # msg = generate_email_notification(users, "weekly cron files copied : {}".format(e))
-            # send_email(msg)
-        else:
-            log_info(
-                f"files already there in folder {daily_cron_file_name1,daily_cron_file_name2}",
-                MODULE_CONTEXT,
-            )
-            msg = generate_email_notification(
-                users, "files already in directory cannot copy"
-            )
-            send_email(msg)
+        )
+
+        saved_docs = stats.fetch_data_for_userwise_trans_user_tokenized(
+            ch_collection, from_date, end_date
+        )
+
+        chdoc = [x for x in ch_docs]
+        ch_doc_df=pd.json_normalize(chdoc)
+        # ch_doc_df.dropna(subset=['orgID'], inplace=True)
+        ch_doc_df.rename(columns = {'created_by':'userID'}, inplace = True)
+        ch_doc_df.reset_index(drop=True, inplace=True)
+        result_ch_doc = user_df.merge(ch_doc_df,indicator=False,how="right")
+        result_ch_doc.to_csv(config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name1)
+
+        savedoc = [x for x in saved_docs]
+        save_doc_df=pd.json_normalize(savedoc)
+        # save_doc_df.dropna(subset=['orgID'], inplace=True)
+        save_doc_df.rename(columns = {'created_by':'userID'}, inplace = True)
+        save_doc_df.reset_index(drop=True, inplace=True)
+        result_save_doc = user_df.merge(save_doc_df,indicator=False,how="right")
+        result_save_doc.to_csv(config.DOWNLOAD_FOLDER + "/" + weekly_cron_file_name2)
+
+        result2 = result_save_doc.merge(result_ch_doc,indicator=False,how="right")
+        result2.to_csv(config.DOWNLOAD_FOLDER + "/" + stats_file)
+
+        log_info(f"Data written into files {weekly_cron_file_name1,weekly_cron_file_name2,stats_file}",MODULE_CONTEXT,)
+
         return
     except Exception as e:
         log_exception("Error in fetching the data: {}".format(e), MODULE_CONTEXT, e)
@@ -147,162 +119,11 @@ def get_trans_user_data_from_db_weekly_crn():
         )
         send_email(msg)
         log_exception(
-            "Generated alert email in exception weekly cron job : {}".format(str(e)),
+            "Generated alert email in exception cron job : {}".format(str(e)),
             MODULE_CONTEXT,
             e,
         )
         return
+    
 
-
-# @schedule_job.scheduled_job(
-#     "cron", id="my_job_id2", day_of_week="mon-fri", hour="00,06,12,18", minute="00"
-# )
-def get_trans_user_data_from_db_daily_day_crn_1():
-    users = config.EMAIL_NOTIFIER
-    log_info("fetch data started", MODULE_CONTEXT)
-    # filename = uuid.uuid4().hex
-    daily_cron_file_name1 = config.DAILY_CRON_FILE_NAME1
-    # daily_cron_file_name2 = config.DAILY_CRON_FILE_NAME2
-    # weekly_cron_file_name1 = config.WEEKLY_CRON_FILE_NAME1
-    # weekly_cron_file_name2 = config.WEEKLY_CRON_FILE_NAME2
-    # file_save = str(filename)[:-10]+'_USER_WISE_JUD_Org_level_Statistics.csv'
-    if os.path.exists(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name1):
-        pass
-    else:
-        msg = generate_email_notification(
-            users, "could not get the data files not found"
-        )
-        send_email(msg)
-        log_info(
-            f"Generated alert email for daily cron (files not found),{daily_cron_file_name1} ",
-            MODULE_CONTEXT,
-        )
-    user_docs = stats.get_all_users_active_inactive(usr_collection)
-    log_info(
-        f"Data returned from user {config.USER_COLLECTION} collection", MODULE_CONTEXT
-    )
-    try:
-        df = pd.read_csv(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name1)
-        from_datee = df["created_on"].max()
-        from_date = datetime.datetime.strptime(str(from_datee), "%Y-%m-%d %H:%M:%S.%f")
-        now = datetime.datetime.now()
-        date_time = now.strftime("%Y-%m-%d")
-        end_date = datetime.datetime.strptime(str(date_time), "%Y-%m-%d")
-        # from_date, end_date = stats.get_time_frame_for_analytics()
-        for doc in user_docs:
-            # log_info(f'fetching details for {doc} userID',MODULE_CONTEXT)
-            ch_docs = stats.fetch_data_for_language_trans_tokenized_for_scheduer_only(
-                ch_collection, doc, from_date, end_date
-            )
-            # saved_docs = stats.fetch_data_for_userwise_trans_user_tokenized(
-            #     ch_collection, doc, from_date, end_date
-            # )
-            # try:
-            #     ch_docs.next()
-            #     saved_docs.next()
-            # except Exception as e:
-            #     print(str(e))
-            # log_info(f'Details collected for for userID : {doc} ',MODULE_CONTEXT)
-            write_to_csv_user_daily_crn(
-                [x for x in ch_docs],
-                (config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name1),
-            )
-            # write_to_csv_user_daily_crn(
-            #     [x for x in saved_docs],
-            #     (config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name2),
-            # )
-        log_info(
-            f"Data written into files {daily_cron_file_name1}",
-            MODULE_CONTEXT,
-        )
-        return
-    except Exception as e:
-        log_exception("Error in fetching the data: {}".format(e), MODULE_CONTEXT, e)
-        msg = generate_email_notification(
-            users, "could not get the data something went wrong : {}".format(e)
-        )
-        send_email(msg)
-        log_exception(
-            "Generated alert email in exception daily cron job : {}".format(str(e)),
-            MODULE_CONTEXT,
-            e,
-        )
-        return
-
-
-# @schedule_job.scheduled_job(
-#     "cron", id="my_job_id3", day_of_week="mon-fri", hour="01,07,13,19", minute="00"
-# )
-def get_trans_user_data_from_db_daily_day_crn_2():
-    users = config.EMAIL_NOTIFIER
-    log_info("fetch data started", MODULE_CONTEXT)
-    # filename = uuid.uuid4().hex
-    daily_cron_file_name1 = config.DAILY_CRON_FILE_NAME1
-    daily_cron_file_name2 = config.DAILY_CRON_FILE_NAME2
-    # weekly_cron_file_name1 = config.WEEKLY_CRON_FILE_NAME1
-    # weekly_cron_file_name2 = config.WEEKLY_CRON_FILE_NAME2
-    # file_save = str(filename)[:-10]+'_USER_WISE_JUD_Org_level_Statistics.csv'
-    if os.path.exists(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name2):
-        pass
-    else:
-        msg = generate_email_notification(
-            users, "could not get the data files not found"
-        )
-        send_email(msg)
-        log_info(
-            f"Generated alert email for daily cron (files not found),{daily_cron_file_name2} ",
-            MODULE_CONTEXT,
-        )
-    user_docs = stats.get_all_users_active_inactive(usr_collection)
-    log_info(
-        f"Data returned from user {config.USER_COLLECTION} collection", MODULE_CONTEXT
-    )
-    try:
-        df = pd.read_csv(config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name1)
-        from_datee = df["created_on"].max()
-        from_date = datetime.datetime.strptime(str(from_datee), "%Y-%m-%d %H:%M:%S.%f")
-        now = datetime.datetime.now()
-        date_time = now.strftime("%Y-%m-%d")
-        end_date = datetime.datetime.strptime(str(date_time), "%Y-%m-%d")
-        # from_date, end_date = stats.get_time_frame_for_analytics()
-        for doc in user_docs:
-            # log_info(f'fetching details for {doc} userID',MODULE_CONTEXT)
-            # ch_docs = stats.fetch_data_for_language_trans_tokenized_for_scheduer_only(
-            #     ch_collection, doc, from_date, end_date
-            # )
-            saved_docs = stats.fetch_data_for_userwise_trans_user_tokenized(
-                ch_collection, doc, from_date, end_date
-            )
-            write_to_csv_user_daily_crn(
-                [x for x in saved_docs],
-                (config.DOWNLOAD_FOLDER + "/" + daily_cron_file_name2),
-            )
-        log_info(
-            f"Data written into files {daily_cron_file_name2}",
-            MODULE_CONTEXT,
-        )
-        return
-    except Exception as e:
-        log_exception("Error in fetching the data: {}".format(e), MODULE_CONTEXT, e)
-        msg = generate_email_notification(
-            users, "could not get the data something went wrong : {}".format(e)
-        )
-        send_email(msg)
-        log_exception(
-            "Generated alert email in exception daily cron job : {}".format(str(e)),
-            MODULE_CONTEXT,
-            e,
-        )
-        return
-
-
-# schedule_job.add_job(
-#     get_trans_user_data_from_db_daily_day_crn,
-#     CronTrigger.from_crontab("49 03,06,12,18 * * *"),
-# )
-# schedule_job.add_job(
-#     get_trans_user_data_from_db_weekly_crn, CronTrigger.from_crontab("00 00 * * 6")
-# )
-
-
-# schedule_job.start()
+schedule_job.start()
